update api authorization permissions

+++ Tabloid/Controllers/CategoryController.cs
@@ -4,6 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Tabloid.Models;
 using Tabloid.Repositories;
@@ -15,34 +16,34 @@ namespace Tabloid.Controllers
     [ApiController]
     public class CategoryController : ControllerBase
     {
-            private readonly IPostRepository _postRepository;
-            private readonly IPostTagRepository _postTagRepository;
-            private readonly ICategoryRepository _categoryRepository;
-            private readonly ICommentRepository _commentRepository;
-            private readonly IUserProfileRepository _userProfileRepository;
+        private readonly IPostRepository _postRepository;
+        private readonly IPostTagRepository _postTagRepository;
+        private readonly ICategoryRepository _categoryRepository;
+        private readonly ICommentRepository _commentRepository;
+        private readonly IUserProfileRepository _userProfileRepository;
 
 
-            public CategoryController(IPostRepository postRepository, IPostTagRepository postTagRepository, ICategoryRepository categoryRepository, ICommentRepository commentRepository, IUserProfileRepository userProfileRepository)
-            {
-                _postRepository = postRepository;
-                _postTagRepository = postTagRepository;
-                _categoryRepository = categoryRepository;
-                _commentRepository = commentRepository;
-                _userProfileRepository = userProfileRepository;
-            }
+        public CategoryController(IPostRepository postRepository, IPostTagRepository postTagRepository, ICategoryRepository categoryRepository, ICommentRepository commentRepository, IUserProfileRepository userProfileRepository)
+        {
+            _postRepository = postRepository;
+            _postTagRepository = postTagRepository;
+            _categoryRepository = categoryRepository;
+            _commentRepository = commentRepository;
+            _userProfileRepository = userProfileRepository;
+        }
 
-            // GET: api/<CategoryController>
-            [HttpGet]
-            public IActionResult Index()
-            {
-                var posts = _categoryRepository.GetAllCategories();
-                return Ok(posts);
-            }
+        // GET: api/<CategoryController>
+        [HttpGet]
+        public IActionResult Index()
+        {
+            var posts = _categoryRepository.GetAllCategories();
+            return Ok(posts);
+        }
 
-            // GET api/<CategoryController>/5
-            [HttpGet("{id}")]
-            public ActionResult Get(int id)
-            {
+        // GET api/<CategoryController>/5
+        [HttpGet("{id}")]
+        public ActionResult Get(int id)
+        {
             var category = _categoryRepository.GetCategoryById(id);
             return Ok(category);
         }
@@ -51,6 +52,11 @@ namespace Tabloid.Controllers
         [HttpPost]
         public IActionResult Post(Category category)
         {
+            if (!IsAllowedAdminPermissions())
+            {
+                return Unauthorized();
+            }
+
             _categoryRepository.CreateCategory(category);
             return NoContent();
         }
@@ -58,14 +64,24 @@ namespace Tabloid.Controllers
         [HttpDelete("{id}")]
         public IActionResult Delete(int id)
         {
+            if (!IsAllowedAdminPermissions())
+            {
+                return Unauthorized();
+            }
+
             _categoryRepository.Delete(id);
             return NoContent();
         }
 
         // PUT api/<CategoryController>/5
         [HttpPut("{id}")]
-            public IActionResult Put(int id, [FromBody] Category category)
+        public IActionResult Put(int id, [FromBody] Category category)
+        {
+            if (!IsAllowedAdminPermissions())
             {
+                return Unauthorized();
+            }
+
             if (id != category.Id)
             {
                 return BadRequest();
@@ -74,6 +90,32 @@ namespace Tabloid.Controllers
             _categoryRepository.Update(category);
             return NoContent();
         }
-            
+        public bool IsAllowedAuthorPermissions()
+        {
+            var firebaseUserId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            var currentUserProfile = _userProfileRepository.GetByFirebaseUserId(firebaseUserId);
+            string currentUserType = currentUserProfile.UserType.Name;
+
+            var globalPermissions = new List<string>()
+            {
+                "admin", "author", "proposed_deactivate", "proposed_demote"
+            };
+
+            return globalPermissions.Contains(currentUserType);
+        }
+
+        public bool IsAllowedAdminPermissions()
+        {
+            var firebaseUserId = User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            var currentUserProfile = _userProfileRepository.GetByFirebaseUserId(firebaseUserId);
+            string currentUserType = currentUserProfile.UserType.Name;
+
+            var adminPermissions = new List<string>()
+            {
+                "admin", "proposed_demote"
+            };
+
+            return adminPermissions.Contains(currentUserType);
         }
     }
+}
